add income module

+++ pkg/httputil/parse.go
@@ -43,13 +43,13 @@ func GetQuerySlice(q url.Values, name string) []string {
 	return strings.Split(val, ",")
 }
 
-func GetQueryInt(q url.Values, name string, defaultVal int64) int64 {
+func GetQueryInt(q url.Values, name string, defaultVal int64) int {
 	val, err := strconv.ParseInt(q.Get(name), 10, 64)
 	if err != nil {
 		val = defaultVal
 	}
 
-	return val
+	return int(val)
 }
 
 func GetQueryBool(q url.Values, name string, defaultVal bool) bool {
